Use streamlit login from PMAT (#182)
