+++ scripts/image_app.py
@@ -1,11 +1,21 @@
+"""
+PYTHONPATH=. streamlit run scripts/image_app.py
+
+Tip: if you specify PYTHONPATH=., streamlit will watch for the changes in all files, instead of just this one.
+"""
+
 import streamlit as st
 from openai import OpenAI
 from prediction_market_agent_tooling.markets.markets import (
     MarketType,
     get_binary_markets,
 )
+from prediction_market_agent_tooling.tools.streamlit_user_login import streamlit_login
 
 st.title("Prediction market thumbnail generator")
+
+with st.sidebar:
+    streamlit_login()
 
 markets = get_binary_markets(42, MarketType.OMEN)
